requirement.txt
flask
pymysql
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
Flask-Uploads
flask-bootstrap
flask-msearch